move components folder and fix build errors
Fixes https://github.com/owncast/owncast/issues/689
