+++ components/config/README.md
