+++ components/config/edit-directory.tsx
@@ -4,8 +4,8 @@ import { Typography } from 'antd';
 
 import ToggleSwitch from './form-toggleswitch-with-submit';
 
-import { ServerStatusContext } from '../../../utils/server-status-context';
-import { FIELD_PROPS_NSFW, FIELD_PROPS_YP } from './constants';
+import { ServerStatusContext } from '../../utils/server-status-context';
+import { FIELD_PROPS_NSFW, FIELD_PROPS_YP } from '../../utils/config-constants';
 
 const { Title } = Typography;
