+++ components/log-table.tsx
