+++ components/message-visiblity-toggle.tsx
@@ -7,10 +7,10 @@ import {
   CheckCircleFilled,
   ExclamationCircleFilled,
 } from '@ant-design/icons';
-import { fetchData, UPDATE_CHAT_MESSGAE_VIZ } from '../../utils/apis';
-import { MessageType } from '../../types/chat';
-import { OUTCOME_TIMEOUT } from '../chat';
-import { isEmptyObject } from '../../utils/format';
+import { fetchData, UPDATE_CHAT_MESSGAE_VIZ } from '../utils/apis';
+import { MessageType } from '../types/chat';
+import { OUTCOME_TIMEOUT } from '../pages/chat';
+import { isEmptyObject } from '../utils/format';
 
 interface MessageToggleProps {
   isVisible: boolean;
