+++ pages/_app.tsx
@@ -23,7 +23,7 @@ import { AppProps } from 'next/app';
 import ServerStatusProvider from '../utils/server-status-context';
 import AlertMessageProvider from '../utils/alert-message-context';
 
-import MainLayout from './components/main-layout';
+import MainLayout from '../components/main-layout';
 
 function App({ Component, pageProps }: AppProps) {
   return (
