+++ pages/config-public-details.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Typography } from 'antd';
 import Link from 'next/link';
 
-import EditInstanceDetails from './components/config/edit-instance-details';
-import EditDirectoryDetails from './components/config/edit-directory';
-import EditInstanceTags from './components/config/edit-tags';
+import EditInstanceDetails from '../components/config/edit-instance-details';
+import EditDirectoryDetails from '../components/config/edit-directory';
+import EditInstanceTags from '../components/config/edit-tags';
 
 const { Title } = Typography;
