+++ pages/config-video.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Typography } from 'antd';
 
-import VideoVariantsTable from './components/config/video-variants-table';
-import VideoLatency from './components/config/video-latency';
+import VideoVariantsTable from '../components/config/video-variants-table';
+import VideoLatency from '../components/config/video-latency';
 
 const { Title } = Typography;
